+++ src/dipdup/config.py
@@ -52,9 +52,12 @@ from dipdup.enums import ReindexingReason
 from dipdup.enums import SkipHistory
 from dipdup.exceptions import ConfigInitializationException
 from dipdup.exceptions import ConfigurationError
+from dipdup.exceptions import ConflictingHooksError
 from dipdup.exceptions import IndexAlreadyExistsError
+from dipdup.exceptions import InitializationRequiredError
 from dipdup.utils import exclude_none
 from dipdup.utils import import_from
+from dipdup.utils import is_importable
 from dipdup.utils import pascal_to_snake
 from dipdup.utils import snake_to_pascal
 
@@ -1225,17 +1228,31 @@ class DipDupConfig:
         except ImportError:
             return os.path.join(os.getcwd(), self.package)
 
+    # TODO: Remove in 6.0
+    @cached_property
+    def per_index_rollback(self) -> bool:
+        """Check if package has `on_index_rollback` hook"""
+        new_hook = is_importable(f'{self.package}.hooks.on_index_rollback', 'on_index_rollback')
+        old_hook = is_importable(f'{self.package}.hooks.on_rollback', 'on_rollback')
+        if new_hook and new_hook:
+            raise ConflictingHooksError('on_rollback', 'on_index_rollback')
+        elif not new_hook and not old_hook:
+            raise InitializationRequiredError('none of `on_rollback` or `on_index_rollback` hooks found')
+        elif new_hook:
+            return True
+        elif old_hook:
+            return False
+        else:
+            raise RuntimeError
+
     @property
     def oneshot(self) -> bool:
         """Whether all indexes have `last_level` field set"""
         syncable_indexes = tuple(c for c in self.indexes.values() if not isinstance(c, HeadIndexConfig))
         oneshot_indexes = tuple(c for c in syncable_indexes if c.last_level)
-        if not oneshot_indexes:
-            return False
-        elif len(oneshot_indexes) == len(syncable_indexes):
+        if len(oneshot_indexes) == len(syncable_indexes):
             return True
-        else:
-            raise ConfigurationError('Either all or none of indexes can have `last_level` field set')
+        return False
 
     @classmethod
     def load(
@@ -1452,29 +1469,25 @@ class DipDupConfig:
             return
 
         if isinstance(index_config, OperationIndexConfig):
-            if self.advanced.merge_subscriptions:
-                index_config.subscriptions.add(TransactionSubscription())
-                return
+            if OperationType.transaction in index_config.types:
+                if self.advanced.merge_subscriptions:
+                    index_config.subscriptions.add(TransactionSubscription())
+                else:
+                    for contract_config in index_config.contracts:
+                        if not isinstance(contract_config, ContractConfig):
+                            raise ConfigInitializationException
+                        index_config.subscriptions.add(TransactionSubscription(address=contract_config.address))
 
-            for contract_config in index_config.contracts:
-                if not isinstance(contract_config, ContractConfig):
-                    raise ConfigInitializationException
-                index_config.subscriptions.add(TransactionSubscription(address=contract_config.address))
-
-            for handler_config in index_config.handlers:
-                for pattern_config in handler_config.pattern:
-                    if isinstance(pattern_config, OperationHandlerOriginationPatternConfig):
-                        index_config.subscriptions.add(OriginationSubscription())
-                        break
+            if OperationType.origination in index_config.types:
+                index_config.subscriptions.add(OriginationSubscription())
 
         elif isinstance(index_config, BigMapIndexConfig):
             if self.advanced.merge_subscriptions:
                 index_config.subscriptions.add(BigMapSubscription())
-                return
-
-            for big_map_handler_config in index_config.handlers:
-                address, path = big_map_handler_config.contract_config.address, big_map_handler_config.path
-                index_config.subscriptions.add(BigMapSubscription(address=address, path=path))
+            else:
+                for big_map_handler_config in index_config.handlers:
+                    address, path = big_map_handler_config.contract_config.address, big_map_handler_config.path
+                    index_config.subscriptions.add(BigMapSubscription(address=address, path=path))
 
         elif isinstance(index_config, HeadIndexConfig):
             index_config.subscriptions.add(HeadSubscription())