+++ src/dipdup/index.py
@@ -201,7 +201,6 @@ class Index:
         )
 
     async def process(self) -> None:
-        # NOTE: `--oneshot` flag implied
         if not isinstance(self._config, HeadIndexConfig) and self._config.last_level:
             head_level = self._config.last_level
             with ExitStack() as stack:
@@ -391,11 +390,13 @@ class OperationIndex(Index):
         if batch_level < index_level:
             raise RuntimeError(f'Batch level is lower than index level: {batch_level} < {index_level}')
 
+        # NOTE: Single level rollback was triggered
         if head_level := self._next_head_level:
             if head_level != index_level:
                 raise RuntimeError(f'New head level is not equal to index level: {head_level} != {index_level}')
 
             self._logger.info('Rolling back to the previous level, verifying processed operations')
+            rollback_hook_called = False
             old_head_hashes = set(self._head_hashes)
             old_head_matched_hashes = {k for k, v in self._head_hashes.items() if v}
             new_head_hashes = {s.hash for s in operation_subgroups}
@@ -405,18 +406,30 @@ class OperationIndex(Index):
 
             self._logger.info('Comparing hashes: %s new, %s missing', len(unprocessed_hashes), len(missing_hashes))
             if missing_hashes:
-                self._logger.info('Some operations were backtracked, requesting reindexing')
-                await self._ctx.reindex(
-                    ReindexingReason.rollback,
-                    datasource=self._datasource.name,
-                    from_level=head_level,
-                    # NOTE: Index level is not decreased on a single-level rollback
-                    to_level=head_level - 1,
-                    missing_hashes=', '.join(missing_hashes),
-                )
+                rollback_hook_called = True
+                self._logger.info('Some operations were backtracked, calling rollback hook')
+                if self._ctx.config.per_index_rollback:
+                    hook_name = 'on_index_rollback'
+                    await self._ctx.fire_hook(
+                        hook_name,
+                        index=self,
+                        from_level=head_level,
+                        to_level=head_level - 1,
+                    )
+                else:
+                    hook_name = 'on_rollback'
+                    await self._ctx.fire_hook(
+                        hook_name,
+                        datasource=self.datasource,
+                        from_level=head_level + 1,
+                        to_level=head_level,
+                    )
 
             self._next_head_level = None
             self._head_hashes.clear()
+
+            if rollback_hook_called:
+                return
 
             operation_subgroups = tuple(filter(lambda subgroup: subgroup.hash in unprocessed_hashes, operation_subgroups))
 
@@ -591,8 +604,10 @@ class OperationIndex(Index):
 
     async def _get_origination_addresses(self) -> Set[str]:
         """Get addresses to fetch origination from during initial synchronization"""
-        # FIXME: Missing `OperationType.origination` in config is ignored
-        addresses = set()
+        if OperationType.origination not in self._config.types:
+            return set()
+
+        addresses: Set[str] = set()
         for handler_config in self._config.handlers:
             for pattern_config in handler_config.pattern:
                 if not isinstance(pattern_config, OperationHandlerOriginationPatternConfig):